refactor: setup DI container inside its own package (fed-sdk)
